+++ template_engines/backends/docx.py
@@ -1,14 +1,13 @@
 from pathlib import Path
 
 from bs4 import BeautifulSoup
-from django.conf import settings
-from django.template.exceptions import TemplateDoesNotExist
 from django.template.context import make_context
+from django.template.exceptions import TemplateDoesNotExist
 
-from . import DOCX_PARAGRAPH_RE, TO_CHANGE_RE, DOCX_CHANGES
-from .abstract import AbstractTemplate, ZipAbstractEngine
-from .utils import modify_content_document
-from .utils_docx import add_image_in_docx_template
+from template_engines import settings as app_settings
+from template_engines.utils import modify_content_document
+from template_engines.utils.docx import add_image_in_docx_template, DOCX_PARAGRAPH_RE, DOCX_CHANGES, TO_CHANGE_RE
+from . import AbstractTemplate, ZipAbstractEngine
 
 
 class DocxTemplate(AbstractTemplate):
@@ -61,8 +60,8 @@ class DocxEngine(ZipAbstractEngine):
     an ``DOCX_ENGINE_SUB_DIRNAME`` setting in your settings.
     By default, ``DocxTemplate`` is used as template_class.
     """
-    sub_dirname = getattr(settings, 'DOCX_ENGINE_SUB_DIRNAME', 'docx')
-    app_dirname = getattr(settings, 'DOCX_ENGINE_APP_DIRNAME', 'templates')
+    sub_dirname = app_settings.DOCX_ENGINE_SUB_DIRNAME
+    app_dirname = app_settings.DOCX_ENGINE_APP_DIRNAME
     template_class = DocxTemplate
     zip_root_files = ['word/document.xml']